Add minimalistic CLI for `MusicPageExporter` to `user_scripts/export_music.py`.
